pyproject.toml
[project]
name = "Opogen"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Equipo45", email = "ismiruiz@gmail.com"},
]
dependencies = [
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "langchain>=0.1.14",
    "langchain-openai>=0.1.1",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
